2nd Audio Jack, Mounting Holes
Moved the Joystick ports to the same side, I'm considering rotating the
whole thing 180 degrees so the joysticks are on the bottom...
Added a second audio jack to the 16 bit Wing.
The first audio jack is now connected to the Analog input pins of the
atmel chip and also to FPGA pins. This is for reading data off tape
drives or recording audio.
Added mounting holes.